refactor: move and rename LearningThemeFilters
