--- src/components/Filters/LearningThemeFilters/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./LearningThemeFilters";
-export type { LearningThemeSelectedTrackingProps } from "./LearningThemeFilters";
