+++ src/components/TeacherComponents/UnitsLearningThemeFilters/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./UnitsLearningThemeFilters";
+export type { LearningThemeSelectedTrackingProps } from "./UnitsLearningThemeFilters";
