+++ src/pages/teachers/programmes/[programmeSlug]/units.tsx
@@ -22,7 +22,7 @@ import curriculumApi, { UnitListingData } from "@/node-lib/curriculum-api";
 import UnitList from "@/components/TeacherComponents/UnitList";
 import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import Box from "@/components/SharedComponents/Box";
-import LearningThemeFilters from "@/components/Filters/LearningThemeFilters";
+import UnitsLearningThemeFilters from "@/components/TeacherComponents/UnitsLearningThemeFilters";
 import MobileFilters from "@/components/MobileFilters";
 import { Heading } from "@/components/SharedComponents/Typography";
 import TabularNav from "@/components/SharedComponents/TabularNav";
@@ -173,7 +173,7 @@ const UnitListingPage: NextPage<UnitListingPageProps> = ({
                     {/* Though still called "Learning themes" internally, these should be referred to as "Threads" in user facing displays */}
                     Filter by thread
                   </Heading>
-                  <LearningThemeFilters
+                  <UnitsLearningThemeFilters
                     labelledBy={learningThemesId}
                     learningThemes={learningThemes}
                     selectedThemeSlug={themeSlug ? themeSlug : "all"}
@@ -217,7 +217,7 @@ const UnitListingPage: NextPage<UnitListingPageProps> = ({
                     $mt={0}
                     $mb={[16, 0]}
                   >
-                    <LearningThemeFilters
+                    <UnitsLearningThemeFilters
                       labelledBy={learningThemesFilterId}
                       learningThemes={learningThemes}
                       selectedThemeSlug={themeSlug ? themeSlug : "all"}
